Add files via upload
Michael Verano Tumala
E-COMMERCE WEBSITE

An e-commerce website is a website through which a business can sell products or services. Once created, it must be so configured that it provides a smooth shopping experience with simple navigation, minimal distractions, and a mobile-friendly design to engage users. 

Our e-commerce platform is a static website that is meant to display products, such as descriptions, prices, and images. This system does not have a working shopping cart or payment gateway like dynamic e-commerce sites. Instead, potential buyers would have to contact the seller directly for orders. Our project's goal is to create a simple yet effective online catalog that provides precise product information and is user-friendly.

A prime illustration of this is Vape Lounge, an internet keep that gives premium puff objects. Some of the choices on the list include BLACK ELITE V1 PODS (8000 puffs), X-VAPE X3, SHIFT DR FREEZE BLACK BAR (9000 puffs) and the alternative BOSS X NEVOKS (19,000 puffs). It’s all part of the user-friendly website where customers can browse and buy their favorite devices. Vape Lounge is committed to providing a hassle-free, no-frills buying experience for vaping enthusiasts; through focusing on high-quality and varied products, they provide an easy way to find just what you are looking for.

We used HTML and CSS to build our website because they're crucial in creating and designing web pages. HTML is used to build the content on the website, such as product lists and navigation menus, while CSS handles the aesthetics, such as colors, layouts, and ensuring that everything appears well on various devices. Flexbox, grid, and media queries are among the features that help the website fit several screen sizes, and remain accessible on a wide range of devices. A good product presentation with visually attractive website design increases the chances of conversion. An appealing and well-designed website improves product browsing and the user experience.

The growing trend in e-commerce, especially purposeful buying, also reflects the need for dynamic e-commerce platforms with real-time update, shopping cart and auto-transaction features like Shopee. Static websites like this project are easier to maintain, load faster, and are less complex but they work well for product presentations that don't require heavy backend functionality. It is important to know these factors when developing an e-commerce website that serves purpose.

Our goal is to set our knowledge of web development to action, implementing this static website with an emphasis on how it looks, how well it is organized and simple to utilize. It cannot process transactions, but it may serve as a learning platform that may one day be fitted with more features. This case is a good example of how access, effectively structured content and transparency in design lead to creating a good presentation of online products